@@ -1,7 +1,7 @@
 {
-  "name": "nypl-app-boilerplate",
-  "version": "0.1.0",
-  "description": "Isomorphic React boilerplate",
+  "name": "nypl-discovery",
+  "version": "0.2.0",
+  "description": "Isomorphic React App for NYPL Research Catalog.",
   "main": "index.js",
   "scripts": {
     "start": "node index",
@@ -51,13 +51,14 @@
   },
   "devDependencies": {
     "axios-mock-adapter": "1.7.1",
+    "create-react-class": "15.5.3",
     "nightwatch": "0.9.9",
     "nyc": "9.0.1"
   },
   "dependencies": {
     "@nypl/design-toolkit": "0.1.16",
-    "@nypl/dgx-header-component": "1.5.0",
-    "@nypl/dgx-react-footer": "0.3.0",
+    "@nypl/dgx-header-component": "2.0.0",
+    "@nypl/dgx-react-footer": "0.4.0",
     "alt": "0.18.2",
     "axios": "0.9.1",
     "babel-core": "6.5.2",
@@ -79,7 +80,7 @@
     "dgx-react-ga": "git+https://git@bitbucket.org/NYPL/dgx-react-ga.git#master",
     "dgx-svg-icons": "git+https://git@bitbucket.org/NYPL/dgx-svg-icons.git#master",
     "ejs": "2.3.3",
-    "enzyme": "2.4.1",
+    "enzyme": "2.8.2",
     "esdoc": "0.4.5",
     "eslint": "2.2.0",
     "eslint-config-airbnb": "13.0.0",
@@ -93,16 +94,17 @@
     "jsonwebtoken": "7.1.9",
     "mocha": "3.0.2",
     "node-sass": "3.8.0",
-    "react": "0.14.8",
+    "prop-types": "15.5.10",
+    "react": "15.5.4",
     "react-a11y": "0.3.3",
     "react-addons-test-utils": "15.3.1",
     "react-doc-meta": "0.2.0",
-    "react-document-title": "2.0.2",
-    "react-dom": "0.14.8",
+    "react-document-title": "2.0.3",
+    "react-dom": "15.5.4",
     "react-ga": "1.2.0",
     "react-hot-loader": "1.2.8",
     "react-onclickout": "2.0.6",
-    "react-router": "2.8.1",
+    "react-router": "3.0.5",
     "react-tappable": "0.8.0",
     "rootpath": "0.1.2",
     "sass-loader": "3.2.0",
