@@ -1,5 +1,5 @@
 import React from 'react';
-
+import PropTypes from 'prop-types';
 // import { SearchIcon } from 'dgx-svg-icons';
 
 const SearchButton = ({
@@ -21,10 +21,10 @@ const SearchButton = ({
 );
 
 SearchButton.propTypes = {
-  onClick: React.PropTypes.func,
-  label: React.PropTypes.string,
-  id: React.PropTypes.string,
-  className: React.PropTypes.string,
+  onClick: PropTypes.func,
+  label: PropTypes.string,
+  id: PropTypes.string,
+  className: PropTypes.string,
 };
 
 SearchButton.defaultProps = {
