@@ -1,5 +1,5 @@
 import React from 'react';
-
+import PropTypes from 'prop-types';
 import DocumentTitle from 'react-document-title';
 
 import Hits from '../Hits/Hits.jsx';
@@ -121,15 +121,15 @@ const SearchResultsPage = (props) => {
 };
 
 SearchResultsPage.propTypes = {
-  searchResults: React.PropTypes.object,
-  searchKeywords: React.PropTypes.string,
-  facets: React.PropTypes.object,
-  selectedFacets: React.PropTypes.object,
-  page: React.PropTypes.string,
-  location: React.PropTypes.object,
-  sortBy: React.PropTypes.string,
-  field: React.PropTypes.string,
-  spinning: React.PropTypes.bool,
+  searchResults: PropTypes.object,
+  searchKeywords: PropTypes.string,
+  facets: PropTypes.object,
+  selectedFacets: PropTypes.object,
+  page: PropTypes.string,
+  location: PropTypes.object,
+  sortBy: PropTypes.string,
+  field: PropTypes.string,
+  spinning: PropTypes.bool,
 };
 
 export default SearchResultsPage;
